Structure fable-core in modules
