--- src/fable/Fable.Core/npm/.babelrc
@@ -1,45 +1,0 @@
-{
-  "env": {
-    "target-umd": {
-      "presets": [
-        "es2015"
-      ],
-      "plugins": [
-        "transform-es2015-modules-umd",
-        [
-          "transform-es2015-classes",
-          {
-            "loose": true
-          }
-        ]
-      ]
-    },
-    "target-commonjs": {
-      "presets": [
-        "es2015"
-      ],
-      "plugins": [
-        "transform-es2015-modules-commonjs",
-        [
-          "transform-es2015-classes",
-          {
-            "loose": true
-          }
-        ]
-      ]
-    },
-    "target-es2015": {
-      "presets": [
-        "es2015-no-commonjs"
-      ],
-      "plugins": [
-        [
-          "transform-es2015-classes",
-          {
-            "loose": true
-          }
-        ]
-      ]
-    }
-  }
-}
