--- src/fable/Fable.Core/npm/.npmignore
@@ -1,2 +1,0 @@
-FSharp.Core.*
-Fable.Core.dll.*
