--- src/fable/Fable.Core/npm/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "es6",
-        "module": "commonjs",
-        "noImplicitAny": true,
-        "noImplicitReturns": true
-    },
-    "files": [
-        "fable-core.ts"
-    ]
-}
