+++ src/fable/Fable.Core/ts/Lazy.ts
@@ -1,0 +1,23 @@
+export function createFromValue<T>(v: T) {
+    return new Lazy(() => v);
+}
+
+export default class Lazy<T> {
+  public factory: () => T;
+  public isValueCreated: boolean;
+
+  private createdValue: T;
+
+  constructor(factory: () => T) {
+    this.factory = factory;
+    this.isValueCreated = false;
+  }
+
+  get value() {
+    if (!this.isValueCreated) {
+      this.createdValue = this.factory();
+      this.isValueCreated = true;
+    }
+    return this.createdValue;
+  }
+}
