+++ src/fable/Fable.Core/ts/Main.ts
@@ -1,0 +1,65 @@
+import * as Array from "./Array"
+export { Array as Array }
+
+import * as Async from "./Async"
+export { Async as Async }
+
+import * as AsyncBuilder from "./AsyncBuilder"
+export { AsyncBuilder as AsyncBuilder }
+
+import * as Choice from "./Choice"
+export { Choice as Choice }
+
+import * as Date from "./Date"
+export { Date as Date }
+
+import * as Event from "./Event"
+export { Event as Event }
+
+import * as GenericComparer from "./GenericComparer"
+export { GenericComparer as GenericComparer }
+
+import * as Lazy from "./Lazy"
+export { Lazy as Lazy }
+
+import * as List from "./List"
+export { List as List }
+
+import * as MailboxProcessor from "./MailboxProcessor"
+export { MailboxProcessor as MailboxProcessor }
+
+import * as Map from "./Map"
+export { Map as Map }
+
+import * as Observable from "./Observable"
+export { Observable as Observable }
+
+import * as Reflection from "./Reflection"
+export { Reflection as Reflection }
+
+import * as RegExp from "./RegExp"
+export { RegExp as RegExp }
+
+import * as Seq from "./Seq"
+export { Seq as Seq }
+
+import * as Serialize from "./Serialize"
+export { Serialize as Serialize }
+
+import * as Set from "./Set"
+export { Set as Set }
+
+import * as String from "./String"
+export { String as String }
+
+import * as Symbol from "./Symbol"
+export { Symbol as Symbol }
+
+import * as Timer from "./Timer"
+export { Timer as Timer }
+
+import * as TimeSpan from "./TimeSpan"
+export { TimeSpan as TimeSpan }
+
+import * as Util from "./Util"
+export { Util as Util }
