+++ src/tests/DllRef/fableconfig.json
@@ -1,5 +1,5 @@
 {
     "projFile": "Fable.Tests.DllRef.fsproj",
     "outDir": "../../../build/tests/DllRef",
-    "module": "commonjs"
+    "coreLib": "../../../build/fable-core"
 }
