+++ src/tests/Other/fableconfig.json
@@ -1,6 +1,6 @@
 {
     "projFile": "Fable.Tests.Clamp.fsproj",
     "outDir": "../../../build/tests/Another Project",
-    "module": "commonjs",
+    "coreLib": "../../../build/fable-core",
     "clamp": true
 }
