+++ src/tests/fableconfig.json
@@ -1,6 +1,7 @@
 {
     "projFile": "Fable.Tests.fsproj",
     "outDir": "../../build/tests",
+    "coreLib": "../../build/fable-core",
     "module": "commonjs",
     "plugins": ["../plugins/nunit/Fable.Plugins.NUnit.dll"],
     "refs": {
