+++ src/tools/QuickTest.fsx
@@ -7,9 +7,9 @@
 // Please don't add this file to your commits
 
 // Uncomment these lines if you need access to Fable.Core helpers:
-// #r "../fable/Fable.Core/npm/Fable.Core.dll"
-// open Fable.Core
-// open Fable.Core.JsInterop
+#r "../../build/fable-core/Fable.Core.dll"
+open Fable.Core
+open Fable.Core.JsInterop
 
 let equal expected actual =
     let areEqual = expected = actual
@@ -19,6 +19,7 @@ let equal expected actual =
 
 // Write here the code you want to test,
 // you can later put the code in a unit test.
+printfn "Hola %s" "Ramón"
 
 // Example:
 // Seq.except [2] [1; 3; 2] |> Seq.last |> equal 3
